dog-pig-loaf-of-bread/requirements.txt
tensorflow>=2.12.0
pillow>=9.0.0
matplotlib>=3.5.0
numpy>=1.21.0
scikit-learn>=1.0.0
